requirements.txt
imagehash
Pillow
python-magic
